feat(zink): add more tests
